Fiz muitas alterações. Criei classes para representar as gomas, o dinheiro, os estado do menu e da maquina. O programa ja reconhece a mudança dos menus de acordo com o estado atual. Ainda falta alguns ajustes.

+++ MaquinaGomas/src/ifes/poo/moeda/Moeda.java
@@ -1,5 +1,5 @@
 package ifes.poo.moeda;
 
 public interface Moeda {
-	public int getValor();
+	public double getValor();
 }
